+++ frontend/package.json
@@ -35,6 +35,7 @@
     "react-icons": "^4.10.1",
     "react-query": "^3.39.3",
     "react-router-dom": "^6.14.2",
+    "react-sketch-canvas": "^6.2.0",
     "typescript": "^4.9.5",
     "web-vitals": "^3.4.0"
   },
