+++ frontend/src/components/canvas/helpers.ts
@@ -1,0 +1,14 @@
+export function hexToRgb(hex: string) {
+  console.log("Hex color: ", hex);
+  if (hex.includes("Nan")) {
+    return null;
+  }
+  var result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex);
+  return result
+    ? {
+        r: parseInt(result[1], 16),
+        g: parseInt(result[2], 16),
+        b: parseInt(result[3], 16),
+      }
+    : null;
+}
